move home page to a specified folder
